pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-rebac"
version = "0.1.0"
description = "RBAC/ReBAC authorization helpers and admin UI for FastAPI applications."
authors = [{ name = "Ldokov Roman" }]
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["fastapi", "rbac", "rebac", "authorization", "access-control"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
]
dependencies = [
    "fastapi>=0.110,<1.0",
    "fastapi-users[sqlalchemy]>=14.0,<15.0",
    "SQLAlchemy>=2.0,<3.0",
    "pydantic>=2.0,<3.0",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0,<3.0",
    "greenlet>=3.0,<4.0",
]

[project.optional-dependencies]
anomaly = [
    "pyod>=2.0,<4.0",
]
postgres = [
    "asyncpg>=0.29,<1.0",
    "alembic>=1.13,<2.0",
]
yandex = [
    "httpx>=0.27,<1.0",
]
example = [
    "asyncpg>=0.29,<1.0",
    "alembic>=1.13,<2.0",
    "python-dotenv>=1.0,<2.0",
    "uvicorn[standard]>=0.27,<1.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.8,<2.0",
]

[project.urls]
Homepage = "https://github.com/lra-itsolution/fastapi-rebac/"
Repository = "https://github.com/lra-itsolution/fastapi-rebac/"
Issues = "https://github.com/lra-itsolution/fastapi-rebac/issues"

[tool.setuptools.packages.find]
include = ["fastapi_rebac*"]
exclude = ["example_app*"]

[tool.setuptools.package-data]
fastapi_rebac = [
    "py.typed",
    "templates/**/*.html",
    "static/**/*",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["fastapi_rebac", "example_app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
